quell warnings
git-svn-id: https://svn.code.sf.net/p/bzflag/code/trunk/bzflag@18849 08b3d480-bf2c-0410-a26f-811ee3361c24

+++ include/NetHandler.h
@@ -120,7 +120,7 @@ public:
   */
   RxStatus    tcpReceive(bool doCodes = true);
   void       *getTcpBuffer();
-  unsigned int getTcpReadSize ( void  );
+  size_t      getTcpReadSize ( void  );
 
   /// Request if there is any buffered udp messages waiting to be sent
   static bool	anyUDPPending() {return pendingUDP;};
